# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Fresh Portal — контракт инженерного пилота R1
  version: 0.1.0-proposed
  description: |
    НА СОГЛАСОВАНИЕ. Contract only, not a deployed API or production readiness claim.
    Narrow R1 slice of TZ v2.10 §§13.1A,45 and security rules. Existing personal
    login/password + server sessions; no new OIDC. Synthetic A/B only.
    17 operations; business mutations require Idempotency-Key for at least 30 days.
    login/logout are the only mutating exceptions. Every action POST explicitly
    carries expected_entity_version; field PATCH uses field CAS instead.
    Check current authorization before schema/CAS and before idempotent replay.
    Hidden object is 404; forbidden explicit branch filter/action is 403.
    Commit mutation + audit + outbox + successful idempotency result atomically.
    All responses include X-Request-ID and Cache-Control:no-store.
    Error envelope code,message,request_id,details also applies to malformed JSON
    (422), gateway errors and quota errors. All timestamps returned in UTC with Z.
    423 is reserved for the general restriction contract; sanctions disabled in R1.
servers:
  - url: /api/v1
...
security:
  - sessionCookie: []
paths:
...
  /notifications:
    get:
      tags: [Notifications]
      operationId: listNotifications
      summary: Только личные доступные IN_APP уведомления
      description: |
        Current recipient identity AND branch grant AND item-read permission.
        Revoked content excluded, no total count. Ascending (created_at,id).
        Clicking a work_item_id uses getWorkItem with fresh authorization.
      x-requirement-ids: [R1-02, R1-10]
      x-permission: notification.read
      x-events: []
      parameters:
        - name: unread_only
          in: query
          schema: {type: boolean, default: false}
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Current authorized inbox page.
          headers:
            X-Request-ID: {$ref: '#/components/headers/RequestId'}
            Cache-Control: {$ref: '#/components/headers/NoStore'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/NotificationPage'}
              example: {items: [], next_cursor: null}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Unprocessable'}
        '423': {$ref: '#/components/responses/Locked'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /notifications/{notification_id}/read:
    parameters:
      - name: notification_id
        in: path
        required: true
        schema: {$ref: '#/components/schemas/Uuid'}
    post:
      tags: [Notifications]
      operationId: readNotification
      summary: CAS отметки о прочтении своего уведомления
      description: |
        Check recipient and current item scope again, including idempotent replay.
        Unknown/other recipient/hidden item is 404. Version CAS applies even when
        already read. First read sets server read_at and version+1 + audit/outbox.
        Already read with current version is 200 no-op; no repeated business event.
      x-requirement-ids: [R1-10, R1-11, R1-12]
      x-permission: notification.read
      x-events: [notification.read]
      security: [{sessionCookie: [], csrfToken: []}]
      parameters:
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/VersionRequest'}
            example: {expected_entity_version: 1}
      responses:
        '200':
          description: Committed read marker or version-checked no-op.
          headers:
            X-Request-ID: {$ref: '#/components/headers/RequestId'}
            Cache-Control: {$ref: '#/components/headers/NoStore'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Notification'}
              example:
                id: '60000000-0000-4000-8000-000000000001'
                event_id: '50000000-0000-4000-8000-000000000002'
                recipient_user_id: '10000000-0000-4000-8000-000000000002'
                org_unit_id: '00000000-0000-4000-8000-00000000000a'
                work_item_id: '30000000-0000-4000-8000-000000000001'
                channel: IN_APP
                message: 'Вам назначена задача.'
                created_at: '2026-09-15T07:42:01Z'
                read_at: '2026-09-15T07:43:00Z'
                entity_version: 2
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Unprocessable'}
        '423': {$ref: '#/components/responses/Locked'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
        '503': {$ref: '#/components/responses/Unavailable'}
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-fresh_session
      description: Existing portal server session, opaque Secure HttpOnly cookie; never a role-bearing JWT.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Session-bound token from login/getMe; required together with cookie on all non-login unsafe operations.
  headers:
    RequestId:
      schema: {type: string, format: uuid}
      description: Server trace UUID; equals error body's request_id if error.
    NoStore:
      schema: {type: string, enum: [no-store]}
      description: Do not cache personal content, tokens or errors.
    RetryAfter:
      schema: {type: integer, minimum: 1}
      description: Seconds before retry, then exponential backoff with jitter.
      example: 30
  parameters:
    Origin:
      name: Origin
      in: header
      required: true
      description: Exact approved HTTPS origin; allowlist hostname TBD before R1 starts. Missing/null/other origin gives 403.
      schema: {type: string, format: uri}
      example: https://pilot.example.invalid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Unique(actor,operation,key); ≥30d. Fingerprint includes target/path and
        canonical JSON payload. Same payload returns saved status/body only after
        fresh read authorization. Changed payload/target ->409. Missing key ->422.
        All business mutations, not login/logout. Do not retry expired create blindly.
      schema: {type: string, minLength: 16, maxLength: 128, pattern: '^[A-Za-z0-9._:-]+$'}
      example: 'pilot-r1-20260915-000001'
...
    Limit:
      name: limit
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 50}
    Cursor:
      name: cursor
      in: query
      description: Opaque actor/filter/permission-bound keyset cursor, proposed TTL15min; not an access capability.
      schema: {type: string, minLength: 1, maxLength: 512}
  schemas:
    Uuid:
      type: string
      format: uuid
    NullableUuid:
      type: string
      format: uuid
      nullable: true
    Version:
      type: integer
      format: int64
      minimum: 1
    UtcTimestamp:
      type: string
      format: date-time
      pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(\.\d{1,6})?Z$'
      description: UTC RFC3339 instant, Z required; semantic calendar validation server-side.
    NullableUtcTimestamp:
      type: string
      format: date-time
      nullable: true
      pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(\.\d{1,6})?Z$'
...
    WorkItemStatus:
      type: string
      enum: [DRAFT, ASSIGNED, IN_PROGRESS, SUBMITTED, COMPLETED, CANCELLED]
...
    VersionRequest:
      type: object
      additionalProperties: false
      required: [expected_entity_version]
      properties:
        expected_entity_version: {$ref: '#/components/schemas/Version'}
...
    NextCursor:
      type: string
      nullable: true
      minLength: 1
      maxLength: 512
...
    Notification:
      type: object
      additionalProperties: false
      required: [id, event_id, recipient_user_id, org_unit_id, work_item_id, channel, message, entity_version, created_at, read_at]
      properties:
        id: {$ref: '#/components/schemas/Uuid'}
        event_id: {$ref: '#/components/schemas/Uuid'}
        recipient_user_id: {$ref: '#/components/schemas/Uuid'}
        org_unit_id: {$ref: '#/components/schemas/Uuid'}
        work_item_id: {$ref: '#/components/schemas/Uuid'}
        channel: {type: string, enum: [IN_APP]}
        message: {type: string, minLength: 1, maxLength: 300, description: 'Safe fixed event text, not user summary or secret data.'}
        entity_version: {$ref: '#/components/schemas/Version'}
        created_at: {$ref: '#/components/schemas/UtcTimestamp'}
        read_at: {$ref: '#/components/schemas/NullableUtcTimestamp'}
    NotificationPage:
      type: object
      additionalProperties: false
      required: [items, next_cursor]
      properties:
        items: {type: array, maxItems: 100, items: {$ref: '#/components/schemas/Notification'}}
        next_cursor: {$ref: '#/components/schemas/NextCursor'}
    FieldConflict:
      type: object
      additionalProperties: false
      required: [field_path, expected_version, current_version, current_value]
      properties:
        field_path: {type: string, enum: [completion_summary]}
        expected_version: {$ref: '#/components/schemas/Version'}
        current_version: {$ref: '#/components/schemas/Version'}
        current_value: {type: string, nullable: true, minLength: 1, maxLength: 4000}
    ValidationIssue:
      type: object
      additionalProperties: false
      required: [path, issue]
      properties:
        path: {type: string, maxLength: 200}
        issue: {type: string, maxLength: 300}
    ErrorDetails:
      type: object
      additionalProperties: false
      description: No hidden org/field values or secrets. Empty object for generic/hidden-object errors.
      properties:
        current_entity_version: {$ref: '#/components/schemas/Version'}
        current_submission_id: {$ref: '#/components/schemas/NullableUuid'}
        current_submission_revision: {type: integer, minimum: 0}
        current_status: {$ref: '#/components/schemas/WorkItemStatus'}
        conflicts:
          type: array
          maxItems: 1
          items: {$ref: '#/components/schemas/FieldConflict'}
        issues:
          type: array
          items: {$ref: '#/components/schemas/ValidationIssue'}
        retry_after_seconds: {type: integer, minimum: 1}
    Error:
      type: object
      additionalProperties: false
      required: [code, message, request_id, details]
      properties:
        code:
          type: string
          enum: [UNAUTHENTICATED, INVALID_CREDENTIALS, SESSION_REVOKED, FORBIDDEN, CSRF_INVALID, ORIGIN_DENIED, SELF_REVIEW_FORBIDDEN, FORBIDDEN_FIELD, NOT_FOUND, ENTITY_VERSION_CONFLICT, FIELD_VERSION_CONFLICT, SUBMISSION_CONFLICT, IDEMPOTENCY_KEY_REUSED, IDEMPOTENCY_IN_PROGRESS, SCHEMA_MISMATCH, VALIDATION_ERROR, INVALID_TRANSITION, COMPLETION_REQUIRED, WORK_ITEM_BLOCKED, ASSIGNEE_INELIGIBLE, INVALID_CURSOR, RESTRICTION_ACTIVE, RATE_LIMITED, TEMPORARILY_UNAVAILABLE]
        message: {type: string, minLength: 1, maxLength: 1000}
        request_id: {$ref: '#/components/schemas/Uuid'}
        details: {$ref: '#/components/schemas/ErrorDetails'}
  responses:
...
    Unauthorized:
      description: 401 absent/expired/revoked session; login uses uniform INVALID_CREDENTIALS.
      headers:
        X-Request-ID: {$ref: '#/components/headers/RequestId'}
        Cache-Control: {$ref: '#/components/headers/NoStore'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
          example: {code: UNAUTHENTICATED, message: 'Требуется персональный вход.', request_id: '70000000-0000-4000-8000-000000000001', details: {}}
    Forbidden:
      description: 403 permission/field/CSRF/Origin/self-review denied; no forbidden data in details.
      headers:
        X-Request-ID: {$ref: '#/components/headers/RequestId'}
        Cache-Control: {$ref: '#/components/headers/NoStore'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
          example: {code: FORBIDDEN, message: 'Действие не разрешено.', request_id: '70000000-0000-4000-8000-000000000002', details: {}}
    NotFound:
      description: 404 missing or hidden object/notification, indistinguishable response.
      headers:
        X-Request-ID: {$ref: '#/components/headers/RequestId'}
        Cache-Control: {$ref: '#/components/headers/NoStore'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
          example: {code: NOT_FOUND, message: 'Объект не найден.', request_id: '70000000-0000-4000-8000-000000000003', details: {}}
    Conflict:
      description: 409 CAS/entity/submission/key/schema conflict; whole request rolled back. Values only after field authorization.
      headers:
        X-Request-ID: {$ref: '#/components/headers/RequestId'}
        Cache-Control: {$ref: '#/components/headers/NoStore'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
          examples:
            field:
              value:
                code: FIELD_VERSION_CONFLICT
                message: 'Поле изменилось, требуется явное разрешение конфликта.'
                request_id: '70000000-0000-4000-8000-000000000004'
                details:
                  current_entity_version: 4
                  conflicts: [{field_path: completion_summary, expected_version: 1, current_version: 2, current_value: 'Сохранённый результат.'}]
            submission:
              value:
                code: SUBMISSION_CONFLICT
                message: 'Версия сдачи устарела.'
                request_id: '70000000-0000-4000-8000-000000000005'
                details: {current_submission_id: '40000000-0000-4000-8000-000000000002', current_submission_revision: 2}
    Unprocessable:
      description: 422 malformed JSON/schema/business guard/invalid cursor; no partial effects.
      headers:
        X-Request-ID: {$ref: '#/components/headers/RequestId'}
        Cache-Control: {$ref: '#/components/headers/NoStore'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
          examples:
            transition:
              value: {code: INVALID_TRANSITION, message: 'Переход из текущего состояния запрещён.', request_id: '70000000-0000-4000-8000-000000000006', details: {current_status: CANCELLED}}
            blocked:
              value: {code: WORK_ITEM_BLOCKED, message: 'Нельзя сдать заблокированную задачу.', request_id: '70000000-0000-4000-8000-000000000007', details: {}}
    Locked:
      description: 423 reserved restriction contract; no sanction activation or legal restriction workflow in R1.
      headers:
        X-Request-ID: {$ref: '#/components/headers/RequestId'}
        Cache-Control: {$ref: '#/components/headers/NoStore'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
          example: {code: RESTRICTION_ACTIVE, message: 'Действует ограничение доступа.', request_id: '70000000-0000-4000-8000-000000000008', details: {}}
    TooManyRequests:
      description: 429 server quota exceeded; Retry-After required, retry with same idempotency key.
      headers:
        X-Request-ID: {$ref: '#/components/headers/RequestId'}
        Cache-Control: {$ref: '#/components/headers/NoStore'}
        Retry-After: {$ref: '#/components/headers/RetryAfter'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
          example: {code: RATE_LIMITED, message: 'Повторите позже.', request_id: '70000000-0000-4000-8000-000000000009', details: {retry_after_seconds: 30}}
    Unavailable:
      description: 503 transient/unconfirmed outcome; Retry-After required, repeat same key to resolve committed result.
      headers:
        X-Request-ID: {$ref: '#/components/headers/RequestId'}
        Cache-Control: {$ref: '#/components/headers/NoStore'}
        Retry-After: {$ref: '#/components/headers/RetryAfter'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
          example: {code: TEMPORARILY_UNAVAILABLE, message: 'Сервис временно недоступен.', request_id: '70000000-0000-4000-8000-000000000010', details: {retry_after_seconds: 30}}
...